Move MigrationPathExtractor to demo module.
