+++ demo-simple/src/main/java/com/novoda/downloadmanager/demo/migration/MigrationPathExtractor.java
@@ -1,4 +1,8 @@
-package com.novoda.downloadmanager;
+package com.novoda.downloadmanager.demo.migration;
+
+import com.novoda.downloadmanager.DownloadBatchId;
+import com.novoda.downloadmanager.FilePath;
+import com.novoda.downloadmanager.FilePathCreator;
 
 import java.io.File;
 
@@ -6,6 +10,7 @@ public final class MigrationPathExtractor {
 
     private static final String PATH_SEPARATOR = File.separator;
     private static final String EMPTY = "";
+    private static final String BASE_PATH = "";
 
     private MigrationPathExtractor() {
         // Uses static utility methods.
@@ -17,7 +22,7 @@ public final class MigrationPathExtractor {
         String fileName = extractFileName(assetPath);
         String absolutePath = basePath + PATH_SEPARATOR + relativePathWithBatchId + fileName;
         String sanitizedAbsolutePath = absolutePath.replaceAll("//", PATH_SEPARATOR);
-        return new LiteFilePath(sanitizedAbsolutePath);
+        return FilePathCreator.create(BASE_PATH, sanitizedAbsolutePath);
     }
 
     private static String extractRelativePath(String basePath, String assetPath) {
